Lots of changes. Should have been smaller commit
Added a m_forks pointer to a malloced pool of pthread_mutex_t which are
all initialized at once to act as the forks.
Added a function to assign each philo two forks from the before
mentioned pool of mutexes.
Added mutexes around the printf statements in each eat, sleep and wait
function so that the prints are at the correct times and we dont try two
printf calls at the same time.
Added a condition that checks if philo is dead or not for the monitor.
The monitor cannot force a thread to join so it needs to wait for it to
finish. We add conditionals to periodically check if the thread should
continue.